feat(lib): Add test_gen implementation
As the primary feature of the package, this adds the `test_gen!` macro
under the `lib.rs` file, as well as adding the `#![no_std]` directive,
due to library not requiring any dependancy on the standard, or standard
dependant libraries.

This simply acts as an initial commit of the functional code of the
library, with the addition of documentation and tests planned for future
commits.