chore(ci): Update CI to be more specific
Due to the fact this library is developed on the stable toolchain,
this means the formatting should only be checked once, for this single
chain.

As such, this commit once again updates the CI configuration for the
repository, in order to limit formatting to a single job, follow-up
checks are dependant on, as it should be a relatively quick check,
able to invalidate follow-up jobs.