chore(ci): Remove Ignored Path
As part of a previous commit, a file-path patten, to ignore the
`Cargo.toml` file, was added under the `paths` header.
This addition as intented to prepare for the implementation of
Continuous Delivery, as it would've executed CI checks, alongside
additional CD check when the file was updated, and this was intended to
prevent the duplication of these checks.

I failed the validate that GitHub actions behave in the way I expected
them to, though, and this setup would not have behaved as I expected,
the ignored path negating the execution of the checks on
pull-requests modifying the `Cargo.toml` file, rather, only doing so if
it was the only file modified...

As such, this will revert this change, with CD checks being constrained
to the publishing of new releases...