feat(lib): Add `ignore` attr branches
Previously, I'd neglected to add branch patterns for branch patterns,
where all the test cases, should be ignored, in the belief these should
not be set up, until they could be set up to run as expected. I've since
found this isn't necessarily the intention behind the use of the `ignore` attribute,
and have since decided to change this stance in favour of supporting
this as a valid pattern for using `test_gen!`.