src/lib.rs: Add UR, AI Case + Annotations
Once again, I failed to account for a potential pattern for a case, in
this case, unique return types for each case, with all cases being
ignored.

On-top of this addition, this also marks the addition
of annotations for the blocks of cases, as well as each unique case,
to improve the ease of identifying the intended behaviour of the cases
they should produce, as an attempt to improve maintainability.