test/tests.rs + src/lib.rs: Add initial tests
In adding the inital substance of the library's implementation,
this will add tests on-top of this, to ensure
every branch of `test_gen!` is usable, and behaving correctly,
according to the tests located in `test/tests.rs`.

Additionally, as `test_gen!` should be capableof building
in a `#![no_std]` environment, this commit will add two minor test cases
to the `lib.rs` file, alongside those of the `tests.rs` file, to
validate this assertion.