refactor: Clean up branch formats + Tests
Upon reconsidering the state of `test_gen!`, I realised its syntax
was a completely inconsistent mess, and could be streamlined
significantly, in particular by making better use of macro syntax...

In particular, in the case of branches with rules applying to all test-cases,
attributes were specified first, followed by the helper function,
with return type last, imitating function definition syntax.
Conversly, within the syntax for specifying test-cases,
this was inverted, with return type specified first,
followed by case arguments, then attributes.

This flip-flopping was a biproduct of `test_gen!` having been created,
during my process of learning the intricacies of Rust's
declarative macro system, and in turn, being unsucessful in grasping how
ambiguity is handled by the system, as should be evident by the new base
pattern definition...

To this end, case specification has been standardised to follow the same
function definition-esque format, as the format for the rules applied to
all test cases, alongside using optional parameters to either side of
the required case parameters, for specifying the case attributes, and
return type, respectively...

Coinciding with my decision to replicate the formatting of function
definitions as closely as possible, this commit will also see the change
from using a comma as the separator between the helper/case args, and
the return type, to instead using a fat arrow...

Whilst a thin arrow would've been a preferable representation, due to
the `$helper` variable, being defined with the `expr` type, this cannot
be followed by a thin arrow, leading to unfortunately falling back on
the fat arrow, even for cases where it isn't required, for the sake of
consistency...

Lastly, due to this commit fundamentally changing the formatting of the
base varient of `test_gen!`, this commit also updates the crate's tests
to pass with this new formatting.