build(Cargo.toml): Update for proc-macro rework
Beginning initial work on a rework of `test_gen` as a proc-macro with
this commit. In this case, that includes bumping the version number
(0.2.0) to indicate the significant deviation from the initial iteration,
removing the `no-std` category, to coincide with the planned removal of
the `#![no_std]` crate attribute, adding the `proc-macro` item,
dependancies for working with proc-macros, as well as semantic tweaks to
the description (updating meaning, tweaking spelling), and improving the
docs URL...

BREAKING CHANGE: The addition of the `proc-macro` item, means only
proc-macros can be exported by the `test_gen` crate now, meaning the
current macro iteration produces a compiler error.