docs(lib): Add include_str! shorthand
As noted in ealier commits, doctests don't run functions with the
`#[test]` annotation, so they've been defined as external implementation
tests for this crate, using the `include_str!` macro to insert them into
documentation as necessary.

As the naming for these files follow a consistent style convention,
this is a quick commit, updating them to use a macro allowing their
inclusion via a shorthand concatonated into the normal file path,
through a helper macro.