test: Update test names inline with new tests
Several of the additional tests added with the proc-macro rework, were
based on existing tests, and as such, they were given names with slight
distinctions, with similarities tying them to the original tests.

This commit will rename these existing tests, to match the naming
schemes of their new counterparts, as well as docs their inclusion
file-paths to ensure the crate builds sucessfully.