tests(lib): Make test helper more concise
I was slightly unhappy with the previous helper design, immediately
converting the parsed string back to a token stream which was then used
as the input for `syn::parse_str`, when an `&str` was passed as one of
the function's parameters. This change should make the helper's
implementation more concise.