refactor(lib): Remove parse_macro_input! use
Whilst most idiomatic examples of `syn` suggest using
`parse_macro_input!` for handling the parsing of initial token streams
in proc-macros, I was somewhat unhappy with this approach, as it clashed
stylistically when paired with a subsequent method chain, and obfuscated
the macro implementation details when handling parsing errors.

As such, this commit updates `test_gen` to instead use the `syn::parse`
function, and the `Result::map_or_else` method, to handle the parsing
and macro expansion.