refactor(lib): Tweak TestHelper Parse method chain (#35)
Whilst working on a previous commit, refactoring the implementation of
the top-level `test_gen` proc-macro function, I was reminded of a method
style which would allow the combinations of the `Option::map` and
`Option::unwrap_or_else` methods, into a single method call, with the
use of `Option::map_or_else`. This is only a minor commit, to implement
this improvement.