refactor(lib): Modify constructor signature
Initially, for the implementation of `ReturnType::try_parse`,
I wanted to mirror the pattern of constructors in the standard library,
returning an `Option` when the provided arguments
which don't produce a valid state for the returned value.

This resulted in an issue of `Result::transpose` being used for both
cases making use of the function call, to make the inner `Result`
the outer layer of the returned value. The return type
also prohibited the functions use with `ParseBuffer::call`.

Consequently, I've decided instead, that having a return type of
`Result<Option<Self>>` also makes sense, as returning `None` works
semantically as being `Ok` as a result of the function call...