docs(lib): Improve maintainer facing docs (#37)
* docs(lib): Improve maintainer facing docs

Looking over the `lib.rs` file, I noticed I'd annotated the fields of
some structs, where the value was being maintained to preserve the span
from the input, but not for most. As such, this commit annotates
these remaining fields, to denote which of these are retained for these
purposes.

Alongside this, there are some minor changes to maintainer facing doc
comments explaining the elements represented by the structs, to better
reflect the representation of these fields, and the use of their
annotations.