style(lib): Add newline before comment
This is awful as an individual commit, but doesn't work with any other
commit... oh well...

With another comment of a similar style to this elsewhere in the file,
I added a newline between the comment and the code above, in an effort
to help the readability of the comment. This change mirrors that
decision here.